Make some performance improvements
Change the jewels string into a set, and use the 'in' keywords instead
of an expression when evaluating whether a stone is a jewel.

According to the Python wiki, the average time complexity of the 'in'
operator, when performed on a set, is O(1) -compared to O(n) for lists.
Both have O(n) worst case. This seems like a more efficient method.
Although the creation of the set is most likely O(n).
Creating a set also has the added benefit of ensuring that the items
are unique.

Also included the use of the 'in' keyword, as it's more Pythonic.

source: https://wiki.python.org/moin/TimeComplexity#set